[SPARK][BUILD] Allow existing spark2 JMH benchmarks to work with either Spark 2 or Spark 3 (#2595)
Moves spark2 JMH tests to spark directory and update jmh.gradle and build.gradle to allow for running on spark3 and spark2
Makes spark 2 and 3 work with Java 8
