+++ build.gradle
@@ -857,7 +857,6 @@ project(':iceberg-spark') {
 }
 
 if (jdkVersion == '8') {
-  apply from: 'jmh.gradle'
 
   project(':iceberg-spark2') {
     configurations.all {
@@ -1234,6 +1233,7 @@ String getJavadocVersion() {
   }
 }
 
+apply from: 'jmh.gradle'
 apply from: 'baseline.gradle'
 apply from: 'deploy.gradle'
 apply from: 'tasks.gradle'
